+++ twiml/voice/verbs/dial.go
@@ -19,7 +19,7 @@ type DialAttributes struct {
 	RecordingStatusCallbackMethod *string `xml:"recordingStatusCallbackMethod,attr,omitempty"`
 	RecordingTrack                *string `xml:"recordingTrack,attr,omitempty"`
 	ReferMethod                   *string `xml:"referMethod,attr,omitempty"`
-	ReferURL                	  *string `xml:"referUrl,attr,omitempty"`
+	ReferURL                      *string `xml:"referUrl,attr,omitempty"`
 	RingTone                      *string `xml:"ringTone,attr,omitempty"`
 	Sequential                    *bool   `xml:"sequential,attr,omitempty"`
 	TimeLimit                     *int    `xml:"timeLimit,attr,omitempty"`