+++ twiml/voice/verbs/nouns/stream.go
@@ -5,8 +5,8 @@ import "encoding/xml"
 type StreamAttributes struct {
 	ConnectorName        *string `xml:"connectorName,attr,omitempty"`
 	Name                 *string `xml:"name,attr,omitempty"`
-	StatusCallback       *string `xml:"status_callback,attr,omitempty"`
-	StatusCallbackMethod *string `xml:"status_callback_method,attr,omitempty"`
+	StatusCallback       *string `xml:"statusCallback,attr,omitempty"`
+	StatusCallbackMethod *string `xml:"statusCallbackMethod,attr,omitempty"`
 	Track                *string `xml:"track,attr,omitempty"`
 	URL                  *string `xml:"url,attr,omitempty"`
 }